requirements.txt
a2wsgi==1.10.10
adbc_driver_manager==1.11.0
adbc_driver_postgresql==1.11.0
adbc_driver_sqlite==1.11.0
aesara==2.9.4
aiocontextvars==0.2.2
aiodns==4.0.0
amdsmi==7.0.2
anywidget==0.10.0
apache_beam==2.72.0
apex==0.9.10dev
apollo_torch==1.0.3
AppKit==0.2.8
argcomplete==3.6.3
argh==0.31.3
asttokens==3.0.1
atheris==3.0.0
authlib==1.6.11
auto_round==0.12.2
av==17.0.0
azure_storage==0.37.0
azureml==0.2.7
beautifulsoup4==4.11.2
billiard==4.2.4
blinker==1.9.0
blobfile==3.2.0
botocore==1.42.90
brotli==1.2.0
brotlicffi==1.2.0.1
cairocffi==1.7.1
Carbon==1.1.10
causal_conv1d==1.6.1
cbor==1.0.0
chardet==7.4.3
checks==0.2
clearml==2.1.5
codecarbon==3.2.6
colorama==0.4.6
comet_ml==3.57.3
compressed_tensors==0.15.0.1
coremltools==9.0
cryptography==46.0.7
ctypes_snappy==1.03
cudf==0.6.1.post1
cupy==14.0.1
cut_cross_entropy==25.1.1
cutlass==0.4.0
cyordereddict==1.0.0
Cython==3.2.4
dagshub==0.7.0
dask_expr==2.0.0
decord==0.6.0
deeplearning==0.0.1
defcon==0.12.2
defusedxml==0.7.1
djangorestframework==3.17.1
dl==0.1.0
docutils==0.22.4
duckdb==1.5.2
dvclive==3.49.0
elasticsearch==9.3.0
email_validator==1.3.0
essentia==2.1b6.dev1389
etcd==2.0.8
eval_type_backport==0.3.1
eventlet==0.41.0
exceptiongroup==1.3.1
expecttest==0.3.0
faiss==1.5.3
fast_hadamard_transform==1.1.0
fastapi_cli==0.0.24
fastmcp==3.2.4
fastparquet==2026.3.0
fbgemm_gpu==1.6.0
fbscribelogger==0.1.7
flask_login==0.6.3
flint==0.0.1
flute==0.2.0
flytekit==1.16.19
foo==.1
Foundation==0.1.0a0.dev1
fqdn==1.5.1
freetype_py==2.5.1
fugashi==1.5.2
g2p_en==2.1.0
galore_torch==1.0
gguf==0.18.0
gitdb_speedups==0.1.0
glyphsLib==6.13.0
gmpy2==2.3.0
gptqmodel==6.0.3
gradio_client==2.4.1
greenlet==3.4.0
grokadamw==0.1.2
grpc_reflection==1.0.0
grpc_tools==1.0.0
gunicorn==25.3.0
gymnasium==1.2.3
h2==4.3.0
h5py==3.7.0
helion==1.0.0
httptools==0.5.0
hypothesis==6.56.3
ibis==3.3.0
importlib_resources==7.1.0
ini2toml==0.15
ipadic==1.0.0
ipykernel==7.2.0
ipyniivue==2.4.4
ipython==8.12.3
ipywidgets==8.1.8
isal==1.8.0
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.10.0
jaxlib==0.10.0
jnius==1.1.0
joblib==1.2.0
joblibspark==0.6.0
JPype1==1.7.0
jsonpath_ng==1.8.0
jsonpointer==3.1.1
kenlm==0.3.0
kerchunk==0.2.10
kernel_abi_check==0.12.2
kombu==5.6.2
kubernetes_asyncio==35.0.1
langchain_aws==1.4.4
langchain_classic==1.0.4
langchain_cohere==0.5.1
langchain_core==1.2.31
langchain_google_vertexai==3.2.2
langchain_huggingface==1.2.2
langchain_mistralai==1.1.2
langchain_ollama==1.1.0
langchain_openai==1.1.14
ldclient==0.0.1
libcst==1.8.6
librosa==0.11.0
liger_kernel==0.7.0
lighteval==0.13.0
load_dotenv==0.1.0
lomo_optim==0.1.1
lxml==6.0.4
lz4==4.4.5
lzmaffi==0.3.0
mambapy==1.2.0
mask==1.0.0
matchpy==0.5.5
megatron==0.5.1
mistral_common==1.11.0
ml_dtypes==0.2.0
modin==0.37.1
monkeytype==23.3.0
moviepy==2.2.1
msgpack_python==0.5.6
mtrand==0.1
multiset==3.2.0
munkres==1.1.4
gpt_lab==0.1.0
natten==0.21.6
nbformat==5.10.4
neptune==1.14.0.post2
nibabel==5.4.2
nltk==3.9.4
num2words==0.5.14
numexpr==2.9.0
objgraph==3.6.2
odfpy==1.4.1
olefile==0.47
onnx_ir==0.2.0
onnxruntime==1.24.4
onnxscript==0.6.2
opencv_python==4.9.0.80
openpyxl==3.1.5
optimum==2.1.0
optree==0.19.0
optuna==4.8.0
orbax==0.1.9
outcome==1.3.0.post0
paramiko==4.0.0
pdfplumber==0.11.9
pexpect==4.9.0
phonemizer==3.3.0
pkg1==0.0.3
polars==1.39.3
polars_st==0.7.0
pox==0.3.7
pretty_midi==0.2.11
pulp==3.3.0
pwlf==2.5.2
py4j==0.10.9.9
py7zr==1.1.0
pyaudio==0.2.14
pycocotools==2.0.11
pycosat==0.6.6
pycryptodome_test_vectors==1.0.22
pyctcdecode==0.5.0
pycute==0.1.0
pydy==0.8.0
pygit2==1.19.2
pyglet==2.1.14
pygraphviz==1.14
PyInstaller==6.19.0
pyinstrument==5.1.2
pymc==5.28.4
pymc3==3.11.6
pynvml==13.0.1
pyodide==0.0.2
pyOpenSSL==26.0.0
pyperf==2.10.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt5_sip==12.13.0
PyQt6==6.11.0
pysat==3.2.2
PySide2==5.15.2.1
PySide6==6.11.0
pyspark==4.1.1
pytesseract==0.3.13
pytest_run_parallel==0.8.2
pytest_subtests==0.15.0
pythainlp==5.3.4
python-dotenv==1.2.2
python_calamine==0.6.2
python_Levenshtein==0.27.3
python_socks==2.8.1
pytidylib==0.3.2
pytorch_lightning==2.6.1
pyxlsb==1.0.10
qtpy==2.4.3
quart_auth==0.11.0
Quartz==0.0.1.dev0
railroad==0.5.0
rarfile==4.2
rdkit==2026.3.1
rediscluster==0.5.3
reportlab==4.4.10
requests_kerberos==0.15.0
Res==0.1.7
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
rfc3987_syntax==1.1.0
rhoknp==1.8.0
rjieba==0.2.0
s3fs==2026.3.0
sacremoses==0.1.1
sage==0.0.0
schedulefree==1.4.1
scipy_doctest==2.2.0
select_backport==0.2
selenium==4.43.0
sentencepiece==0.2.1
setuptools_scm==10.0.5
shap==0.51.0
shiboken2==5.15.2.1
shiboken6==6.11.0
simpleaudio==1.0.4
simplejson==3.20.2
sip==6.15.3
skimage==0.0
slack_sdk==3.41.0
smart_open==7.6.0
smbprotocol==1.16.1
sniffio==1.3.0
socksio==1.0.0
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.14
speedups==2.1.0
Sphinx==9.1.0
sphinx_gallery==0.20.0
spqr_quant==0.2.0
sqlframe==4.1.0
stable_baselines3==2.8.0
sudachipy==0.6.11
swanlab==0.7.15
symengine==0.14.1
tabulate==0.10.0
tenacity==9.1.4
tensorboard_plugin_profile==2.22.2
tensorboard_plugin_wit==1.8.1
tensorboardX==2.6.5
tensorflow_io==0.37.1
theano==1.0.5
threadpoolctl==3.6.0
timm==1.0.26
toml==0.10.2
tomli_w==1.2.0
torch_neuronx==1.0
torch_npu==2.9.0
torch_xla==2.9.0
torchaudio==2.11.0
torchcodec==0.11.1
torchrec==1.6.0
torchvision==0.26.0
traitlets==5.14.3
trove_classifiers==2026.1.14.14
Twisted==25.5.0
ufoLib2==0.18.1
uharfbuzz==0.53.6
ujson==5.5.0
unicodedata2==17.0.1
unidic==1.1.0
unidic_lite==1.0.8
UnleashClient==6.7.0
uri_template==1.3.0
uroman==1.3.1.1
uwsgi==2.0.31
vegafusion==2.0.3
wandb_workspaces==0.3.9
watchdog==6.0.0
watchfiles==1.1.1
webcolors==25.10.0
WebOb==1.8.9
winloop==0.6.2
wsproto==1.3.2
xarray==2026.4.0
xattr==1.3.0
xlrd==2.0.2
xlsxwriter==3.2.9
xmlrunner==1.7.7
xx==3.3.2
xxhash==3.5.0
yt_dlp==2026.3.17
z3==0.2.0
zope==6.0
zopfli==0.4.1
zstandard==0.25.0
zstd==1.5.7.3
